+++ frontend/src/lib/utils.ts
@@ -24,6 +24,7 @@ export const viewableTools = [
   "morpho-viewer-tool",
   "plot-generator",
   "random-plot-generator",
+  "scsgetone-plot",
 ];
 
 /**